% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_main.R
\name{setDefault}
\alias{setDefault}
\title{Define the default value for a function's argument}
\usage{
setDefault(funName, ...)
}
\arguments{
\item{funName}{Name of the function}

\item{...}{default value assignments in the form of \code{key = value} pairs}
}
\description{
Define the default value for a function's argument
}
\examples{
\dontrun{

# This will lead to an error if funtion "hello" is not defined
setDefault("hello", firstName = "Peter")

# Define the function and use getDefault instead of a constant default value
hello <- function(
  firstName = getDefault("hello", "firstName"),
  lastName = getDefault("hello", "lastName")
)
{
  cat(paste0("Hello ", firstName, " ", lastName, "!\\n") )
}

# Now you can define the argument defaults
setDefault("hello", firstName = "Don", lastName = "Quichote")

# If you call the function without arguments, the defaults are used
hello()

# You can now change the defaults without changing the function definition
setDefault("hello", firstName = "Mona", lastName = "Lisa")

hello()
}
}
\seealso{
\code{\link{getDefault}}
}
