% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_main.R
\name{getDefault}
\alias{getDefault}
\title{Get the default value that is defined for a function's argument}
\usage{
getDefault(funName, argName, default = NULL, warn = TRUE)
}
\arguments{
\item{funName}{Name of the function}

\item{argName}{Name of the formal argument}

\item{default}{Default value to use if no default value is stored for
argument \code{argName} of function \code{funName}}

\item{warn}{if \code{TRUE} (default) a message is given if no defaults are
defined for the given function or argument and if \code{default} is
returned instead.}
}
\value{
default value that is defined for the formal argument \code{argName}
  of the user-defined function \code{funName}
}
\description{
Get the default value that is defined for a function's argument
}
\examples{
# Once you have defined a function...
hello <- function(firstName = getDefault("hello", "firstName")) {
  cat("Hello", firstName, "\\n")
}

# ... you can define the default value for its arguments...
setDefault("hello", firstName = "Peter")

# ... and read it back with getDefault()...
getDefault("hello", "firstName")
}
\seealso{
\code{\link{setDefault}}
}
